Crate cfx_vm_types

Structs

ActionParams

Action (call/create) input params. Everything else should be specified in Externalities.

CIP645Spec

Represents the feature flags for CIP-645 implementation.

ConsensusGasSpec

Spec parameters are determined solely by block height and thus accessible to the consensus protocol.

Env

Information concerning the execution environment for a message-call/contract-creation.

ReturnData

Return data buffer. Holds memory from a previous call and a slice into that memory.

Spec

Definition of the cost spec and other parameterisations for the VM.

Enums

ActionValue

Transaction value

BlockHashSource

CallType

The type of the call-like instruction.

ContractCreateResult

Result of externalities create function.

CreateContractAddress

Specifies how an address is calculated for a new contract.

CreateType

The type of the create-like instruction.

Error

VM errors.

GasLeft

Gas Left: either it is a known value, or it needs to be computed by processing a return instruction.

InstructionResult

MessageCallResult

Result of externalities call function.

ParamsType

Type of the way parameters encoded

TrapError

TrapKind

TrapResult

Constants

CODE_PREFIX_7702

Traits

Context

Context for VMs

Exec

Virtual Machine interface

InterpreterInfo

ResumeCall

Resume call interface

ResumeCreate

Resume create interface

Functions

contract_address

Calculate new contract address.

extract_7702_payload

separate_out_db_error

Separate out database-related errors from other EVM errors. The EVM itself does not distinguish between errors that originate from within (e.g., index out-of-bounds) and those that come from external sources like database operations.

Type Aliases

ExecTrapError

ExecTrapResult

Result