cfx_vm_types

Enum Error

pub enum Error {

    OutOfGas,
    BadJumpDestination {
        destination: usize,
    },
    BadInstruction {
        instruction: u8,
    },
    StackUnderflow {
        instruction: &'static str,
        wanted: usize,
        on_stack: usize,
    },
    OutOfStack {
        instruction: &'static str,
        wanted: usize,
        limit: usize,
    },
    SubStackUnderflow {
        wanted: usize,
        on_stack: usize,
    },
    OutOfSubStack {
        wanted: usize,
        limit: usize,
    },
    InvalidSubEntry,
    NotEnoughBalanceForStorage {
        required: U256,
        got: U256,
    },
    ExceedStorageLimit,
    BuiltIn(String),
    InternalContract(String),
    MutableCallInStaticContext,
    CreateInitCodeSizeLimit,
    StateDbError(PartialEqWrapper<Error>),
    Wasm(String),
    OutOfBounds,
    Reverted,
    InvalidAddress(Address),
    ConflictAddress(Address),
    NonceOverflow(Address),
    CreateContractStartingWithEF,

}

VM errors.

Variants

OutOfGas

OutOfGas is returned when transaction execution runs out of gas. The state should be reverted to the state from before the transaction execution. But it does not mean that transaction was invalid. Balance still should be transfered and nonce should be increased.

BadJumpDestination

BadJumpDestination is returned when execution tried to move to position that wasn’t marked with JUMPDEST instruction

Fields

destination: usize

Position the code tried to jump to.

BadInstruction

BadInstructions is returned when given instruction is not supported

Fields

instruction: u8

Unrecognized opcode

StackUnderflow

StackUnderflow when there is not enough stack elements to execute instruction

Fields

instruction: &'static str

Invoked instruction

wanted: usize

How many stack elements was requested by instruction

on_stack: usize

How many elements were on stack

OutOfStack

When execution would exceed defined Stack Limit

Fields

instruction: &'static str

Invoked instruction

wanted: usize

How many stack elements instruction wanted to push

limit: usize

What was the stack limit

SubStackUnderflow

SubStackUnderflow when there is not enough stack elements to execute a subroutine return

Fields

wanted: usize

How many stack elements was requested by instruction

on_stack: usize

How many elements were on stack

OutOfSubStack

When execution would exceed defined subroutine Stack Limit

Fields

wanted: usize

How many stack elements instruction wanted to pop

limit: usize

What was the stack limit

InvalidSubEntry

NotEnoughBalanceForStorage

When balance is not enough for collateral_for_storage. The state should be reverted to the state from before the transaction execution.

Fields

required: U256

got: U256

ExceedStorageLimit

ExceedStorageLimit is returned when the collateral_for_storage exceed the storage_limit.

BuiltIn(String)

Built-in contract failed on given input

InternalContract(String)

Internal contract failed

MutableCallInStaticContext

When execution tries to modify the state in static context

CreateInitCodeSizeLimit

Too large init code size (CIP-645i: EIP-3860)

StateDbError(PartialEqWrapper<Error>)

Error from storage.

Wasm(String)

Wasm runtime error

OutOfBounds

Out of bounds access in RETURNDATACOPY.

Reverted

Execution has been reverted with REVERT.

InvalidAddress(Address)

Invalid address

ConflictAddress(Address)

Create a contract on an address with existing contract

NonceOverflow(Address)

Create a contract on an address with existing contract

CreateContractStartingWithEF

CIP-150: Reject new contract code starting with the 0xEF byte