cfx_vm_types

Trait Exec

pub trait Exec: Send {
    // Required method
    fn exec(
        self: Box<Self>,
        context: &mut dyn Context,
    ) -> ExecTrapResult<GasLeft>;
}

Virtual Machine interface

Required Methods

fn exec(self: Box<Self>, context: &mut dyn Context) -> ExecTrapResult<GasLeft>

This function should be used to execute transaction. It returns either an error, a known amount of gas left, or parameters to be used to compute the final gas left.